link directly to docs page
Currently the link to the LinkML site brings you to a [version / status page](https://linkml.io/linkml-model/) that primarily serves to redirect users (on click) to the documentation site. The status page currently shows a failed build.

This change shortcuts the intermediate page and links directly to the documentation site.