Change linter CLI exit code based on number of errors and warnings
If no errors or warnings are found, the exit code is 0. If any errors are found, the exit code is 2. If warnings are found, but no errors, the exit code is 1 unless other options are specified.

If there are only warning and the --ignore-warnings flag is provided, the exit code will be 0. If there are only warnings and the --max-warnings option is provided and the number of warnings is less than or equal to the provided number, the exit code will be zero.