Use multi-stage Docker build
This brings the Docker image more inline with what you would get by
installing the LinkML package from PyPI. It uses a multi-stage build to
first do a `poetry build` in one stage. Then a second stage copies
the wheel file from the first and installs it. This means the final image
is cleaner, without a copy of the source code or Poetry installed. Also
the final image uses a non-root user as a matter of good general
practice.