fix(owlgen): no triple needed for imported terms
It is not possible to inherit from terms from external vocabularies (for
example, you cannot write "is_a: schema:Person"), since those terms are
not available in the schema.

One way to make them available in the schema is explicitly declaring
them like this (for the above example with "schema:Person":
    Person:
      class_uri: schema:Person
but it creates an unexpected class declaration in the OWL ontology
saying that "schema:Person is an exactMatch of schema:Person".

This patch gets rid of all those superfluous declarations.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>