sqlalchemygen: resolve id column types
If a table has an id column of a type other than text, a FK column would not be able to link to it, as the generator would set the FK type to text. This change will now look up the real type and set the fk column to that type.