jsonschemagen: use anyOf rather than oneOf for range class descendants
This prevents an error if json data satisfies the schema of more than one range class descendants.
This now happens more than it should because the type designator values are not yet part of the json schema, but it can still happen in other cases if there is no type designator field.