feat(converters): add option include descendants
"gen-json-schema" and "linkml-validate" have the option
"--include-range-class-descendants" so that children of a class
specified in a range are added to the generated JSON-Schema as allowed
classes.

This patch adds a similar option to "linkml-convert", with the same
result in the JSON-Schema being used for validation as the equally named
option for "gen-json-schema" and "linkml-validate".

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>