linkml-lint: ignore dot files by default
Motivation:

The `.yml` / `.yaml` file extension indicates the contents is YAML. YAML
is a common format and is used by some non-LinkML tools as their
configuration file format.

A common convention on Unix-like systems is that configuration files
start with a dot (`.`) character.  Therefore it is relatively common for
applications to ignore or hide files that start with a dot when working
with the contents of a directory.

Currently, when scanning a directory, linkml-lint will attempt to
process all files that have `.yaml` or `.yml` as the file extension.
This includes dot files.

While linkml-lint will avoid processing its own configuration file, it
will attempt to process any YAML configuration files of other tools.
This leads to the tool reporting problems even when the actual LinkML
files are fine.

Modification:

Update `linkml-lint` so that, by default, it ignores files that start
with a dot.  Introduce a new command-line option (`-a` or `--all`) that
overrides this behaviour and includes dot files when scanning a
directory.  In either case, linkml-lint avoids its own configuration
files (if present).

Result:

When scanning for files in a directory, `linkml-lint` now ignores
filenames that start with a dot unless a new command-line option (`-a`
or `--all`) is specified.

Closes: #1827
Signed-off-by: Paul Millar <paul.millar@desy.de>