gen-json-schema: support taking JSON Schema `title` from `title` slot (#1838)
Motivation:

JSON Schema supports `title` and `description` as a schema annotations,
with the following description:

    Both of these keywords can be used to decorate a user interface with
    information about the data produced by this user interface. A title
    will preferably be short, whereas a description will provide
    explanation about the purpose of the instance described by this
    schema.

The nature of these annotations isn't specified; however, I'm assuming
both JSON Schema annotations are usually natural language prose, with
title following common conventions of the title of other creative works
(such as books).

Currently gen-json-schema takes the LinkML schema's name slot as the
corresponding JSON Schema's title annotation.

To me, this is not a perfect mapping, as a "name" (as with a LinkML
schema-level slot) is an identifier and often just a single word, while
a "title" is often a short descriptive phrase and typically includes
multiple words. This distinction is, perhaps, reflected in the
restrictions placed on the name slot:

    name – the schema name. Use only alphanumeric characters,
    underscores, and dashes

More specifically, a JSON Schema title value allows spaces while a
LinkML name slot does not.

The same argument applies for the JSON Schema `title` annotations
generated from LinkML classes and enums.  These currently use the class
or enum name (respectively), although using these names as the title may
be sub-optimal.

The JSON Schema generated for slots currently lacks any `title`
annotation.

Modification:

Update gen-json-schema to accept an extra configuration option
(`--title-from`) that controls how `title` annotations are generated.
This takes two values `name` and `title`, with `name` being the default.

The `name` option provides backwards compatibility.  The generated
output is identical to that produced before this patch.

The `title` option uses the LinkML `title` slot for the JSON Schema
`title` annotation if the slot is present, otherwise it behaves as if
the `name` option was specified.  This is true for JSON Schema generated
from the LinkML Schema slots, classes, enums and slots.

Result:

The gen-json-schema command now includes the `--title-from` option that
controls the support for generating JSON Schema `title` annotations from
the LinkML schema `title` slot.

Closes: #1834

Signed-off-by: Paul Millar <paul.millar@desy.de>