validator: support including validator context when reporting errors (#2089)
* validator: support including validator context when reporting errors

Motivation:

When validating JSON, the `linkml-validate` command builds a JSON Schema
against which the JSON input is checked.  If there are validation
problems, the command prints a single line for each such problem.

In some cases, the failing node is a composition (`allOf`, `anyOf`,
`oneOf`) and the resulting error message is currently rather vague.

Modification:

Update the `ValidationResult` class to include a List of strings.  These
are the validation errors that led to the composition failing.

Note, these contextual failures could (themselves) have further
contextual failures, forming a tree.  In the interests of keeping the
patch small, this change records only the next level; however, a future
patch may address this issue.

The error output is updated to support providing the validation error's
context as lines starting `[CONTEXT]`.

The command-line arguments are updated to support the `-v`/`--verbose`
flag.  Specifying this flag results in the validator printing any
validation error context.

Note, support for the `-v` flag was included for backwards
compatibility; however, it may be acceptable to print the validation
error's context unconditionally.  If so, a future patch could remove
this flag.

Result:

`linkml-validate` now supports the `-v`/`--verbose` option.  Specifying
this option will result in certain validation error reports containing
more information.

Closes: #2088

* Update cli.py

* Update cli.py

---------

Co-authored-by: Chris Mungall <cjm@berkeleybop.org>