Fix compliance tests for shacl:
In b415224a821699c6bb35a1375da1869ac476b517 a change was made to add
'sh:nodeKind sh:Literal' to a property shape for a slot with a type
as range. This required an update to the expected shacl output used
in 'test_compliance/test_core_compliance.py::test_cardinality'.