`genshacl`: fix the propagation of `types` to SHACL `sh:datatype` and `sh:nodeKind` (#2102)
* Fixes the propagation of types to shacl export

In the pervious state, the edited code was never reached. With this
change, a linkml or custom type as slot range will propagate to the
property's sh:datatype in SHACL. In addition, a nodekind of sh:Literal
is added for such properties.

* Fix compliance tests for shacl:

In b415224a821699c6bb35a1375da1869ac476b517 a change was made to add
'sh:nodeKind sh:Literal' to a property shape for a slot with a type
as range. This required an update to the expected shacl output used
in 'test_compliance/test_core_compliance.py::test_cardinality'.