chore: improve json schema for uris and curies
Make JSON-Schema generation for URIs and CURIEs stricter by specifying
the "string" format "uri" [1]. Since CURIEs (in contrast with SafeCURIEs
[2]) are valid URIs, we can use this format.

Side-remark: SafeCURIEs are not valid URIs, but there is no space for
ambiguity (is it a CURIE or an URI?). Should be change somewhen in the
future to SafeCURIEs, we would need to change the generated JSON-Schema
whenever CURIEs are involved.

[1]: https://json-schema.org/understanding-json-schema/reference/string#resource-identifiers
[2]: https://www.w3.org/TR/2010/NOTE-curie-20101216/#P_safe_curie

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>