schema validator plugin.
This is currently incomplete, and subject to change.
The plugin mechanism may not be the best place for this.

Ultimately the schema validator can include logical checks,
e.g. testing for *satisiability* of schemas.

This is also required for testing metamodel *implementation*
slots (but the two are not yet hooked up). Normally a schema is
hardwired to be an instance of a metamodel schema. Using implementation
slots allows the schema author to provide their own metaclasses
effetcively further constraining the base metamodel