[docgen] optional rendering of imported elements
Merging imports and rendering documentation of imports are two different
things, but docgen appears to missuse imports merging to have a somehow
complete/consistent view on all the elements relevant for the schema.

This patch enables the possibility of showing imported elements without
the need to merge the imports. That way it is also possible to show
information about which schema an element comes from.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>