feat(graphql): add support for enumerations (#2303)
* feat(graphql): add support for enumerations

Generation of GraphQL enumerations based on those enumerations defined
in the LinkML schema was missing.

LinkML supports enumeration names and permissible values in a format
that is not compabtible with GraphQL. This patch adds the needed
conversions to make them compabible. A dedicated class is added to
handle it in a generic way that can be reused in other parts of LinkML.

Automatic conversion can be disabled with the flag `--strict`.

Enumerations not using `permissible_values` but `reachable_from` to
specify the enumeration permissible values are not supported yet by the
generator. A warning is raised and the enumeration is ignored if this
happens.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>

* test(gen-graphql): add tests

Add tests for the generation of GraphQL schemas.

Adapt also existing tests afected by changes in gen-graphql.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>

---------

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>