fix(jsonldcontextgen): avoid shadowing dataclass fields
Assignation of non-None default values to "end_schema" is shadowing the
values assigned to the corresponding dataclass fields in the
constructor. This patch fixes it.

After this fix values assigned in the dataclass constructor will be
taken, unless different values are provided to "end_schema" directly.
Otherwise default values are used.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>