Update shaclgen.py
This commit should fix issue #2084, in that it allows to have freely defined shacl_shape_uri's cia annotations['shacl_uri'] and by switching from the previously implemented version with class_uri+suffix to reusing the native linkML class name as the name for the SHACL shape. This should be more aligned to the ideas behind SHACL, where a shape is its own rdf instance.