refactor(loader): report issues as warnings
Report issues in combination of presence of identifier in range class
and values of `inlined` and `inlined_as_list` as warnings.

These changes don't have any external impact (apart from the additional
output). This is a preparation for later setting default values in a
posterior commit, what requires changes in the tests.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>