fix(loader): inline as list by def if no ident
Inlined classes without an identifier are typically inlined as a list,
because it is unclear what to use as key. This patch sets that default.

This applies to the cases where `inlined: True` has been explicitly set.
But also to those were it has been automatically set, because no value
was assigned and inlining was the only possiblity.

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>

fix(loader): inline as list by def if no ident

Signed-off-by: Silvano Cirujano Cuesta <silvano.cirujano-cuesta@siemens.com>