Interface-preserving indexing updates in sqltablegen.py
Mapping the `--autogenerate-index` command-line argument to the `SqlTableGenerator` instance properties `autogenerate_fk_index` and `autogenerate_pk_index` without changing the external interface of `linkml.generators.sqltablegen:cli`. This change was to address `gen-sqltables` breakage.  The "right" way to do this would be to modify the command line arguments, but I do not know what would break. Note I am unsure if the FK behavior is correct, just took a guess.