fix(golanggen): fix missing time import
This commit resolves a bug in the Golang template generator where
missing imports for time-related functionality were encountered when
using Jinja2 templates. The root cause was due to the inability of `set`
in Jinja2 blocks (e.g., loops, conditionals) to reassign outer
variables; it creates variables scoped only to the block.

Signed-off-by: Michael Adler <michael.adler@siemens.com>