pydanticgen: make the default for optional list configurable
Historical behaviour for optional lists in pydantic was to initialise them with None. However, https://github.com/linkml/linkml/pull/2877 changed
this behaviour to empty list. This caused regressions for us (see example below). This patch makes the behaviour of pydantic for optional lists configurable,
so we can restore the old behaviour.

Test case:

```
  AssociatedNetElement:
    description: >-
      A RSM Associated Net Element
    attributes:
      netElement:
        range: PositioningNetElement
        inlined: false
      bounds:
        range: float
        multivalued: true
        minimum_cardinality: 1
        maximum_cardinality: 2
      appliesInDirection:
        range: string
      lateralPosition:
        range: string # FIXME custom class LateralPosition instead of str in RSM
```

After PR #2877 AssociatedNetElement() will raise an error (validation fails because altough `bounds` is not required, when it is filled in, the cardinality is mandatory).