Add Archive Repo Notice
Due to some excellent work over in Data Vocabs we can now detach the
need for this forked dep and archive it. Woop

I've left a notice for any future archeologists trying to figure out why
we did this. And also to hihglight there was never anything sensitive in
here.

At some point we might consider deleting it. It has very little
historical significance - and was a fix for a problem that no longer
exists